fixed spin panel nfms null
